Move all UI code under ui package
