+++ CovidWatch/app/src/main/java/org/covidwatch/android/ble/BluetoothManager.kt
@@ -11,7 +11,7 @@ import android.os.IBinder
 import android.util.Log
 import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat.getSystemService
-import org.covidwatch.android.MainActivity
+import org.covidwatch.android.ui.MainActivity
 import org.covidwatch.android.R
 import org.covidwatch.android.data.ContactEvent
 import org.covidwatch.android.data.ContactEventDAO
