+++ CovidWatch/app/src/main/java/org/covidwatch/android/ui/contactevents/adapters/FragmentBindingAdapters.kt
@@ -1,4 +1,4 @@
-package org.covidwatch.android.adapters
+package org.covidwatch.android.ui.contactevents.adapters
 
 import androidx.fragment.app.Fragment
