+++ CovidWatch/app/src/main/java/org/covidwatch/android/ui/contactevents/adapters/FragmentDataBindingComponent.kt
@@ -1,4 +1,4 @@
-package org.covidwatch.android.adapters
+package org.covidwatch.android.ui.contactevents.adapters
 
 import androidx.databinding.DataBindingComponent
 import androidx.fragment.app.Fragment
