+++ CovidWatch/app/src/main/java/org/covidwatch/android/ui/userprofile/UserProfileFragment.kt
@@ -1,4 +1,4 @@
-package org.covidwatch.android.tagsUi.userprofile
+package org.covidwatch.android.ui.userprofile
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -11,7 +11,7 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import org.covidwatch.android.R
-import org.covidwatch.android.adapters.FragmentDataBindingComponent
+import org.covidwatch.android.ui.contactevents.adapters.FragmentDataBindingComponent
 import org.covidwatch.android.databinding.FragmentUserProfileBinding
