+++ CovidWatch/app/src/main/res/layout/fragment_contact_events.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".tagsUi.contactevents.ContactEventsFragment">
+        tools:context=".ui.contactevents.ContactEventsFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/contact_events_recyclerview"
