+++ CovidWatch/app/src/main/res/layout/fragment_user_profile.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".tagsUi.userprofile.UserProfileFragment"
+        tools:context=".ui.userprofile.UserProfileFragment"
         android:background="@{viewModel.isCurrentUserSick ? @color/colorIsCurrentUserSick : @color/colorIsCurrentUserNotSick }" >
 
         <LinearLayout
@@ -54,7 +54,7 @@
         <import type="org.covidwatch.android.R" />
         <variable
             name="viewModel"
-            type="org.covidwatch.android.tagsUi.userprofile.UserProfileViewModel" />
+            type="org.covidwatch.android.ui.userprofile.UserProfileViewModel" />
     </data>
 
 </layout>
