--- CovidWatch/app/src/main/res/navigation/mobile_navigation.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_contact_events">
-
-    <fragment
-        android:id="@+id/navigation_contact_events"
-        android:name="org.covidwatch.android.tagsUi.contactevents.ContactEventsFragment"
-        android:label="@string/title_contact_events"
-        tools:layout="@layout/fragment_contact_events" />
-
-    <fragment
-        android:id="@+id/navigation_user_profile"
-        android:name="org.covidwatch.android.tagsUi.userprofile.UserProfileFragment"
-        android:label="@string/title_my_profile"
-        tools:layout="@layout/fragment_user_profile" />
-
-</navigation>
